Create Adjunct CmdStan Model Environment

Creates an environment to estimate the \(p \times i\) interaction model with CmdStan.

CreateMod(
  coef_hyper,
  sd_hyper,
  nWarmup_init,
  nSamples_init,
  nWarmup_run,
  nSamples_run,
  aux_envir
)

Arguments

coef_hyper

Hyperparameter value for the standard deviation of normally distributed parameters

sd_hyper

Hyperparameter value for the shape parameter of gamma distributed parameters

nWarmup_init

Number of burn-in draws for the fixed \(\theta\) model

nWarmup_run

Number of burn-in draws for the free \(\theta\) model

nSamples_run

Number of sampled posterior values for the free \(\theta\) model (after burn-in has completed)

aux_envir

An environment object that contains objects to be loaded into the CreateMod environment (such as is returned by the genData function)

nSampels_init

Number of posterior draws for the fixed \(\theta\) model (after burn-in has completed)

Value

an environment object containing fixed \(\theta\) and free \(\theta\) model results

Details

Creates an environment with methods for fitting the fixed \(\theta\) and free \(\theta\) models. Additionally, CreateMod includes a method for recursively checking \(\hat{R}\) convergence.

See also