Data Generating Function for the \(p \times i\) Interaction Model

Generates data for a \(p \times i\) Interaction Model under the Explanatory Item Response theoretical framework.

genData(P, I, J, K, seed = NULL, isCorrI = TRUE)

Arguments

P

Number of examinees

I

Number of items

J

Number of potential item feature categories

K

Number of potential person feature categories

seed

Integer seed for replication (if NULL, a random seed will be generated)

Value

an environment containing simulated data

Details

Let a model of the log-odds transformed outcome \((\eta_{pi})\) be defined as: \[\eta_{pi} = \theta_{p}\lambda_{i} + \tau_{i} + \sum_{k=1}^{K}\sum_{j=1}^{J}\beta_{jk_{(\eta)}}x_{ik}z_{pj}\] where the linear predictor of latent trait measurements \(\theta_{p}\)is defined as: \[\theta_{p} = \sum_{j=1}^{J}\beta_{j_{(\theta)}}z_{pj} + u_{p_{(\theta)}}\] where the linear predictor of item slope/discrimination terms (\(\lambda_{i}\)) is defined as: \[\lambda_{i} = \sum_{k=1}^{K}\beta_{k_{(\lambda)}}x_{ik} + u_{i_{(\lambda)}}\] and where the linear predictor of item intercept/easiness terms (\(\tau_{i}\)) is defined as: \[\tau_{i} = \sum_{k=1}^{K}\beta_{k_{(\tau)}}x_{ik} + u_{i_{(\tau)}}\] For the linear predictor of latent trait measurements, the following assumption is made on its residual terms for the purposes of model identification: \[u_{p_{(\theta)}} \sim \mathcal{N}(0,1)\] If an assumption is made that there is no strcutured dependency between the residual terms for the linear predictors of item intercepts and item slopes (isCorrI = FALSE), then: \[\begin{gathered} u_{i_{(\lambda)}} \sim \mathcal{N}(0, \sigma^{2}_{\lambda}) \cr u_{i_{(\tau)}} \sim \mathcal{N}(0, \sigma^{2}_{\tau})\end{gathered}\] Whereas if it is assumed that the residuals of the linear predictors of item intercepts and item slopes is assumed to have a predictable underlying structure (isCorrI = TRUE), then: \[\begin{bmatrix} u_{i_{(\lambda)}} \cr u_{i_{(\tau)}} \end{bmatrix} \sim \mathcal{MVN} \Bigg(\mu = \begin{bmatrix} 0 \cr 0 \end{bmatrix} , \hspace{2pt} \mathbf{\Sigma} = \begin{bmatrix}\sigma^{2}_{\lambda} & \sigma_{\lambda}\sigma_{\tau} \cr \sigma_{\tau}\sigma_{\lambda} & \sigma^{2}_{\tau} \end{bmatrix} \Bigg)\]