Randomly Sample a Cholesky Factorized, Lower Triangular Matrix from a Lewandowski-Kurowicka-Joe Distribution

Generates a lower triangular, invertible, positive-definite matrix

rCorr(nDim, eta = 1)

Arguments

nDim

desired dimension of the sampled \(n \times n\) matrix

eta

concentration hyperparameter

Value

a lower triangular \(n \times n\) matrix

Details

Using the onion method (Ghosh & Henderson, 2003), samples a Cholesky factorized lower-triangular matrix \(L\) from a Lewandowski-Kurowicka-Joe (LKJ) distribution given the concentration hyperparameter \(\eta\), such that: \[A=LL^{\top}\] where \(A\) is a \(n \times n\) matrix and is a member of the set of all symmetric, positive-definite matrices.

References

Ghosh, S. and Henderson, S. G. (2003). Behavior of the norta method for correlated random vector generation as the dimension increases. ACM Transactions on Modeling and Computer Simulation, 13(3), 276-294. https://doi.org/10.1145/937332.937336